+++ src/pages/NotFound/NotFound.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './NotFound.module.css';
+import styles from './css/NotFound.module.css';
 import { RiEmotionSadLine } from 'react-icons/ri';
 import { Link } from 'react-router-dom';
 
@@ -8,7 +8,7 @@ export default function NotFound() {
     <section className={styles['not-found']}>
       <RiEmotionSadLine className={styles.icon} />
       <p className={styles.message}>페이지를 찾을 수 없습니다.</p>
-      <Link to="/" className={styles.btn} type="button">
+      <Link to='/' className={styles.btn} type='button'>
         홈으로 돌아가기
       </Link>
     </section>
